rct_platform/microservices/vector-search/requirements.txt
# ALGO-16 Vector Search Service Dependencies

# Core
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-multipart==0.0.21

# Vector backends
faiss-cpu==1.9.0
qdrant-client==1.12.1
numpy==2.2.1

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1

# Utilities
python-dotenv==1.0.0
